docs: add Face Discoloration Fix
Normally, Face Discoloration Fix is not required if the mod
conflicts were properly resolved. Although all the conflicts have
been proerly resolved in this mod list, Lawless was an exception
that caused the discoloration issue by messing with vanilla
facegens. I am adding this mod as a workaround since there was no
simple fix for Lawless.